Refactoring: Extracting the client to another module
